requirements.txt
black
pytest==7.2.0
gensim==4.2.0
requests==2.28.1
tqdm==4.64.1
mecab-python3==1.0.6
